x/ride/types/message_finish_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/smarshall-spitzbart/ride/testutil/sample"
	"github.com/stretchr/testify/require"
)

func TestMsgFinish_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgFinish
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgFinish{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgFinish{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
